refactor(core): reorganize model specific files in site-scons
[no changelog]
