+++ core/site_scons/models/T1B1/trezor_1.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
-from . import get_hw_model_as_number
-from .stm32f4_common import stm32f4_common_files
+from .. import get_hw_model_as_number
+from ..stm32f4_common import stm32f4_common_files
 
 
 def configure(
@@ -12,49 +12,33 @@ def configure(
     paths: list[str],
 ) -> list[str]:
     features_available: list[str] = []
-    hw_model = get_hw_model_as_number("T2B1")
-    hw_revision = 4
-    board = "trezor_r_v4.h"
+    board = "trezor_1.h"
     display = "vg-2864ksweg01.c"
+    hw_model = get_hw_model_as_number("T1B1")
+    hw_revision = 0
 
-    if "new_rendering" in features_wanted:
-        defines += ["XFRAMEBUFFER"]
-        features_available.append("xframebuffer")
-        features_available.append("display_mono")
-
-    mcu = "STM32F427xx"
+    mcu = "STM32F405xx"
 
     stm32f4_common_files(env, defines, sources, paths)
 
-    env.get("ENV")[
-        "CPU_ASFLAGS"
-    ] = "-mthumb -mcpu=cortex-m4 -mfloat-abi=hard -mfpu=fpv4-sp-d16"
+    env.get("ENV")["CPU_ASFLAGS"] = "-mthumb -mcpu=cortex-m3 -mfloat-abi=soft"
     env.get("ENV")[
         "CPU_CCFLAGS"
-    ] = "-mthumb -mcpu=cortex-m4 -mfloat-abi=hard -mfpu=fpv4-sp-d16 -mtune=cortex-m4 "
-    env.get("ENV")["RUST_TARGET"] = "thumbv7em-none-eabihf"
+    ] = "-mthumb -mtune=cortex-m3 -mcpu=cortex-m3 -mfloat-abi=soft "
+    env.get("ENV")["RUST_TARGET"] = "thumbv7m-none-eabi"
 
     defines += [mcu]
     defines += [f'TREZOR_BOARD=\\"boards/{board}\\"']
     defines += [f"HW_MODEL={hw_model}"]
     defines += [f"HW_REVISION={hw_revision}"]
     sources += [
-        "embed/models/model_T2B1_layout.c",
+        "embed/models/model_T1B1_layout.c",
     ]
-
-    if "new_rendering" in features_wanted:
-        sources += ["embed/trezorhal/xdisplay_legacy.c"]
-        sources += ["embed/trezorhal/stm32f4/xdisplay/vg-2864/display_driver.c"]
-    else:
-        sources += [f"embed/trezorhal/stm32f4/displays/{display}"]
+    sources += [f"embed/trezorhal/stm32f4/displays/{display}"]
 
     if "input" in features_wanted:
         sources += ["embed/trezorhal/stm32f4/button.c"]
         features_available.append("button")
-
-    if "sbu" in features_wanted:
-        sources += ["embed/trezorhal/stm32f4/sbu.c"]
-        features_available.append("sbu")
 
     if "usb" in features_wanted:
         sources += [
