+++ core/site_scons/models/stm32f4_common.py
