+++ core/site_scons/models/stm32u5_common.py
